big bug

+++ tomo_io.cpp
@@ -105,8 +105,8 @@ int main(int argc, char *argv[])
     }
     env_init(np, rank, P, S, C, write_flag);
     memory_allocation(P, S/np, C);        // allocate memory to emulate output data access
-    int stride = sizeof(float) * C * C;
-    offset = rank * (S/np) * stride;
+    long long stride = sizeof(float) * C * C;
+    offset = (long long)rank * (S/np) * stride;
     count = (S/np) * C * C;
     struct timeval fwrite_s, fwrite_e;
     double fwrite_eps;
